+++ packages/client/src/pages/ForumPage/ForumPage.tsx
@@ -1,172 +1,14 @@
-import { FC, useState } from "react";
-import { Link } from "react-router-dom";
-import { useFormik } from "formik";
-import {
-  Box,
-  CardActionArea, CardHeader,
-  IconButton, InputAdornment,
-  List,
-  Paper, TablePagination,
-  TextField,
-  Typography
-} from "@mui/material";
-import { TopicItem } from "../../features/TopicItem/TopicItem";
-import { searchValidation } from "../../shared/constants/validationShemas";
-import { forumApi } from "../../shared/constants/mocks";
+import { FC } from "react";
+import { Container } from "@mui/material";
 import { MainLayout } from "../../shared/layouts/MainLayout";
-import { ArrowForward, Search } from "@mui/icons-material";
-import { AddTopic } from "../../features/AddTopic/AddTopic";
-import TablePaginationActions from "@mui/material/TablePagination/TablePaginationActions";
+import { Forum } from "../../widgets/Forum/Forum";
 
 export const ForumPage: FC = () => {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(1);
-  const rowsOffset = page * rowsPerPage;
-  const filtredTopics = forumApi.getTopic().slice(rowsOffset, rowsOffset + rowsPerPage);
-
-  const formikSearch = useFormik({
-    initialValues: {
-      search: ""
-    },
-    validationSchema: searchValidation,
-    onSubmit: values => {
-      console.log(values);
-    }
-  });
-  const onChangePage = (
-    event: React.MouseEvent<HTMLButtonElement> | null,
-    newPage: number,
-  ) => {
-    setPage(newPage);
-  };
-
-  const onChangeRowsPerPage = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
-
   return (
     <MainLayout>
-      <Paper
-        variant="outlined"
-        sx={{
-          opacity: ".8",
-          display: "flex",
-          flexDirection: "column",
-          // justifyContent: "center",
-          alignItems: "center",
-          height: "80%",
-          width: "80%",
-          margin: "auto",
-          padding: 3
-        }}>
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            width: "100%"
-          }}>
-          <Typography
-            variant="h2"
-            component="h1"
-            className="topic-page__name"
-            m={"auto"}>
-            Forum
-          </Typography>
-        </Box>
-        <Box
-          sx={{
-            boxSizing: "border-box",
-            display: "flex",
-            alignItems: "center",
-            width: "100%",
-            justifyContent: "space-between",
-            px: 4
-          }}>
-          <Box
-            component="form"
-            onSubmit={formikSearch.handleSubmit}
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "end",
-              flex: 1,
-              my: 2,
-              mr: 5
-            }}>
-            <TextField
-              fullWidth
-              id="search"
-              name="search"
-              label="Search"
-              onChange={formikSearch.handleChange}
-              error={formikSearch.touched.search && Boolean(formikSearch.errors.search)}
-              helperText={formikSearch.touched.search && formikSearch.errors.search}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <Search />
-                  </InputAdornment>
-                ),
-                endAdornment: (
-                  <InputAdornment position="end">
-                    <IconButton
-                      aria-label="search topic"
-                      edge="end"
-                      type="submit"
-                    >
-                      <ArrowForward/>
-                    </IconButton>
-                  </InputAdornment>
-                )
-              }}
-            />
-          </Box>
-          <AddTopic/>
-        </Box>
-        <List
-          sx={{
-            flex: 1,
-            width: "100%",
-            overflowY: "auto"
-          }}>
-          {filtredTopics.map(item => (
-            <TopicItem key={item.id} isBordered {...item} header={() => (
-              <CardActionArea component={Link} to={`/forum/${item.id}`}>
-                <CardHeader
-                  title={<Typography variant="h5" component="h2">Topic {item.id}</Typography>}
-                />
-              </CardActionArea>
-            )} />
-          ))}
-        </List>
-        <Box
-          sx={{
-            display: "flex",
-            justifyContent: "flex-end",
-            width: "100%"
-          }}>
-          <TablePagination
-            component="div"
-            rowsPerPageOptions={[1, 2, 3]}
-            colSpan={4}
-            count={forumApi.getTopic().length}
-            rowsPerPage={rowsPerPage}
-            page={page}
-            SelectProps={{
-              inputProps: {
-                "aria-label": "rows per page",
-              },
-              native: true,
-            }}
-            onPageChange={onChangePage}
-            onRowsPerPageChange={onChangeRowsPerPage}
-            ActionsComponent={TablePaginationActions}
-          />
-        </Box>
-      </Paper>
+      <Container sx={{ mt: "9.375%", mb: "6.45%" }}>
+        <Forum />
+      </Container>
     </MainLayout>
   );
 };
